refactor: decouple op definitions from backend implementations
Separate the op registration system so that op definition files are
backend-agnostic and each backend registers its implementations lazily
through its own registration module.

Key changes:
- Add composed_impl to Op class for backend-agnostic fallback dispatch
- Extract all HLO implementations into _hlo_impls.py with lazy
  registration via _register_hlo.py (same pattern as kernelgen)
- Op definition files now only declare Op instances, CPU impls, and
  composed_impl for ops built from other dispatched ops
- Remove redundant per-backend registration of composed ops from
  _register_kernelgen.py (they fall through to composed_impl)

Adding a new backend now only requires registering primitive ops;
all composed ops (floor_divide, tan, mean, cumsum, etc.) automatically
work via the composed_impl fallback.
